// angularpi/src/app/core/interfaces/dashboard-metrics.interface.ts
export interface DashboardMetrics {
    totalEstudiantes: number;
    totalAlertas: number;
    totalIntervenciones: number;

    riesgoAlto: number;
    riesgoMedio: number;
    riesgoBajo: number;

    promedioAcademico: number;
    promedioAsistencia: number;
    promedioReprobadas: number;

    sinInternet: number;
    victimasViolencia: number;
    trabajoInfantil: number;
}